rename Hub fragment
To allow for more settings in that fragment
