+++ app/src/main/res/navigation/main_navigation.xml
@@ -16,9 +16,9 @@
     </fragment>
     <fragment
         android:id="@+id/navigation_hub"
-        android:name="com.helio.app.ui.hub.HubFragment"
+        android:name="com.helio.app.ui.settings.SettingsFragment"
         android:label="@string/title_connect_hub"
-        tools:layout="@layout/fragment_hub" />
+        tools:layout="@layout/fragment_settings" />
     <fragment
         android:id="@+id/navigation_control"
         android:name="com.helio.app.ui.control.ControlFragment"
