+++ app/src/main/res/values/strings.xml
@@ -20,4 +20,11 @@
     <string name="add_blinds_button">Add blinds</string>
     <string name="title_theme_selection">App Theme</string>
     <string name="summary_theme_selection">Select your favourite theme for the Helio App</string>
+    <string name="enter_your_hub_s_ip_address_to_get_started">Enter your Hub\'s IP address to get started:</string>
+    <string name="connect_to_your_helio_hub_device">Connect to your Helio Hub device</string>
+    <string name="_255">255</string>
+    <string name="connect">Connect</string>
+    <string name="connection_status">Connection status</string>
+    <string name="not_connected">not connected</string>
+    <string name="_dot">.</string>
 </resources>
